requirements.txt
mkdocs-awesome-pages-plugin
mkdocs-simple-hooks
lxml